#1 Feat : Advertising_Exposure : deductChargeForChargeType 함수 실행 시 Advertising의 트랜잭션 중복 접근으로 인한 오류 해결 / 제거 대상과 수정 대상 분리 / 테스트 수정

+++ advertising_exposure/src/main/kotlin/com/advertising/advertising_exposure/controller/dto/AdvertisementRes.kt
@@ -27,7 +27,7 @@ data class AdvertisementRes(
 
         fun fromDocument(advertisementDocument: AdvertisementDocument) =
             AdvertisementRes(
-                advertisementDocument.id,
+                advertisementDocument.id.toLong(),
                 advertisementDocument.shopId,
                 advertisementDocument.image,
                 advertisementDocument.description,
